move plusGradientBrush to Pocket Casts Colors
