+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/PocketCastsColors.kt
@@ -1,5 +1,6 @@
 package au.com.shiftyjelly.pocketcasts.compose
 
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 
 val Color.Companion.pocketRed get() = PocketCastsColors.pocketRed
@@ -9,6 +10,7 @@ val Color.Companion.plusGoldDark get() = PocketCastsColors.plusGoldDark
 val Color.Companion.patronPurple get() = PocketCastsColors.patronPurple
 val Color.Companion.patronPurpleLight get() = PocketCastsColors.patronPurpleLight
 val Color.Companion.patronPurpleDark get() = PocketCastsColors.patronPurpleDark
+val Brush.Companion.plusGradientBrush get() = PocketCastsColors.plusGradientBrush
 
 private object PocketCastsColors {
     val pocketRed = Color(0xFFF43E37)
@@ -18,4 +20,8 @@ private object PocketCastsColors {
     val patronPurple = Color(0xFF6046F5)
     val patronPurpleLight = Color(0xFFAFA2FA)
     val patronPurpleDark = patronPurple
+    val plusGradientBrush = Brush.horizontalGradient(
+        0f to Color.plusGoldLight,
+        1f to Color.plusGoldDark,
+    )
 }
